feat: improve facets a little bit
