+++ keep-ui/app/alerts/alert-table-facet-types.tsx
@@ -1,0 +1,43 @@
+import { AlertDto, Severity } from "./models";
+import { Table } from "@tanstack/table-core";
+import { DynamicFacet } from "./alert-table-facet-dynamic";
+
+export interface FacetValue {
+  label: string;
+  count: number;
+  isSelected: boolean;
+}
+
+export interface FacetFilters {
+  [key: string]: string[];
+}
+
+export interface FacetValueProps {
+  label: string;
+  count: number;
+  isSelected: boolean;
+  onSelect: (value: string, exclusive: boolean, isAllOnly: boolean) => void;
+  facetKey: string;
+  showIcon?: boolean;
+  facetFilters: FacetFilters;
+}
+
+export interface FacetProps {
+  name: string;
+  values: FacetValue[];
+  onSelect: (value: string, exclusive: boolean, isAllOnly: boolean) => void;
+  facetKey: string;
+  facetFilters: FacetFilters;
+  showIcon?: boolean;
+}
+
+export interface AlertFacetsProps {
+  alerts: AlertDto[];
+  facetFilters: FacetFilters;
+  setFacetFilters: (filters: FacetFilters) => void;
+  dynamicFacets: DynamicFacet[];
+  setDynamicFacets: (facets: DynamicFacet[]) => void;
+  onDelete: (facetKey: string) => void;
+  className?: string;
+  table: Table<AlertDto>;
+}
